backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.4
psycopg2-binary==2.9.9
clickhouse-driver==0.2.6
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
openpyxl==3.1.2
streamlit==1.28.2
redis==5.0.1
celery==5.3.4
apache-airflow==2.7.3
requests==2.31.0
